3- Ahmad Al-Zein/requirements.txt
pyaudio
pvporcupine
SpeechRecognition
pygame
google-generativeai
PySide6
spacy
gtts
requests
python-dotenv
comtypes
pycaw